feat: Complete teacher quiz functionality and S85 schema migration
- Add teacher quiz creation with password support
- Implement teacher quiz filtering and dedicated API
- Complete S85 backward compatibility cleanup
- Reorganize quiz functions for better maintainability
- Update database schema documentation
